refactor(cgi): rename ZealPHP\Legacy\FastCgiClient → ZealPHP\CGI\FastCgiClient
The "Legacy" namespace was misleading. The class is OpenSwoole-coroutine-
native at the socket layer (uses OpenSwoole\Coroutine\Client — every
send()/recv() yields under the event loop). Only the FCGI 1.0 record
framing is hand-written PHP; the I/O is fully async. It belongs as a
sibling of ZealPHP\CGI\WorkerPool and ZealPHP\CGI\IPC, not in a
"Legacy" bucket that implies blocking or deprecated code.

Why no upstream-native swap (was P1.9 in v0.3.0 roadmap):
Verified against installed ext-openswoole 26.2 — the full
OpenSwoole\Coroutine\* surface is Channel, Client, Context, Http\Client,
Http2\Client, Iterator, PostgreSQL, Scheduler, Socket, System.
OpenSwoole\Coroutine\FastCGI\Client does NOT exist. The roadmap entry
conflated OpenSwoole with upstream Swoole (Swoole 5.x ships the class;
OpenSwoole forked before that addition and never ported it).

When/if OpenSwoole adds the class upstream, ZealPHP\CGI\FastCgiClient
becomes a thin proxy and ~260 LOC drops to ~30 LOC of adapter — no
user-facing API change.

Files touched (10):
  - src/Legacy/FastCgiClient.php   → src/CGI/FastCgiClient.php (git mv)
                                     namespace ZealPHP\Legacy → ZealPHP\CGI
                                     docblock updated to call out OpenSwoole-
                                     native socket layer + upstream context
  - src/App.php                    cgiFcgi(): new \ZealPHP\CGI\FastCgiClient
                                   + \ZealPHP\CGI\FastCgiException catch
                                   + docblock note about OpenSwoole socket
  - tests/Unit/CgiFcgiDispatchTest.php       use ZealPHP\CGI\FastCgi*
  - tests/Unit/FastCgiClientTest.php         use ZealPHP\CGI\FastCgi*
  - tests/Unit/CgiBackendDispatchTest.php    use ZealPHP\CGI\FastCgiClient
  - docs/fastcgi-backends.md       2 refs updated + drive-by: drop stale
                                   cgiMode('fork') mention (removed in
                                   v0.2.41), point to cgiMode('pool') as
                                   the warm-pool option
  - docs/architecture/2026-05-23-v0.3.0-roadmap.md
                                   P1.9 rewritten as PART-DONE + reality
                                   check on OpenSwoole 26.2 surface
  - docs/CHANGES-SINCE-v0.2.38.md  P1.9 table row updated to match
  - template/pages/legacy-apps.php Under-the-hood code ref updated

Verification:
  - 2912 unit tests pass (up from 2906 — gained Mode 6 + 5 superglobal-
    reach tests in the prior commit; this rename adds no tests)
  - PHPStan level 10 src/ scope: zero errors
  - No stale ZealPHP\Legacy\FastCgi references remain in code or docs
    (the 2 mentions left in docs/architecture/*roadmap.md and
    CHANGES-SINCE-v0.2.38.md are intentional historical context
    describing the rename itself)
