+++ docs/architecture/2026-05-23-v0.3.0-roadmap.md
@@ -187,18 +187,23 @@ $values = Cache::pipeline(function ($p) use ($keys) {
 
 Same shape on `Store::pipeline()`. ~80 LOC + tests + perf bench.
 
-### P1.9 — CGI native path — retire `Legacy\FastCgiClient`
+### P1.9 — CGI FastCGI client — namespace cleanup (status: PART-DONE; native swap BLOCKED upstream)
 
-**State today:** `src/Legacy/FastCgiClient.php` (~260 LOC) is a hand-rolled FastCGI client. OpenSwoole 22.x ships `OpenSwoole\Coroutine\FastCGI\Client` as a built-in.
+**Original target (revised based on reality):** Replace the hand-rolled `Legacy\FastCgiClient` with `OpenSwoole\Coroutine\FastCGI\Client`.
 
-**Target:** Replace the hand-rolled client with the native one. The cgiFcgi dispatch path becomes:
+**Reality check (verified 2026-05-24 against installed OpenSwoole 26.2):** `OpenSwoole\Coroutine\FastCGI\Client` does NOT exist in the installed extension. Full surface under `OpenSwoole\Coroutine\*` is: `Channel, Client, Context, Http\Client, Http2\Client, Iterator, PostgreSQL, Scheduler, Socket, System`. Upstream Swoole 5.x ships `Swoole\Coroutine\FastCGI\Client` but OpenSwoole forked before that addition and never ported it. The original roadmap entry conflated the two extensions.
+
+**What landed instead — namespace cleanup:** `src/Legacy/FastCgiClient.php` moved to `src/CGI/FastCgiClient.php` (namespace `ZealPHP\Legacy` → `ZealPHP\CGI`). The "Legacy" framing was misleading anyway — the socket layer is OpenSwoole-native (`OpenSwoole\Coroutine\Client`), only the FCGI 1.0 record framing is hand-written. Sibling namespace to `ZealPHP\CGI\WorkerPool` / `ZealPHP\CGI\IPC` now. Class API unchanged. ~260 LOC of well-tested protocol code stays in tree.
+
+**Future swap (deferred, blocked on upstream):**
 
 ```php
+// If OpenSwoole eventually adds the class:
 $client = new \OpenSwoole\Coroutine\FastCGI\Client('127.0.0.1', 9000);
 $response = $client->request($params, $stdinBody, $timeoutMs);
 ```
 
-Drop ~260 LOC of FastCGI parser code. Verify httpoxy CVE behaviour (`HTTP_PROXY` strip), 60s timeout, stderr drain still hold. ~30 LOC of replacement + integration test against the existing `cgi-bin` recipes.
+When/if that lands upstream, `ZealPHP\CGI\FastCgiClient` becomes a thin proxy and ~260 LOC drops to ~30 LOC of adapter + the existing httpoxy CVE strip / 60s timeout / stderr drain tests still apply.
 
 ### P1.10 — `App::stats()` aggregate + `/healthz` default ✅ PART-SHIPPED (`bfab6f4` — `App::stats()` done; `/healthz` middleware + Prometheus exposition still TBD)
