+++ tests/Unit/CgiBackendDispatchTest.php
@@ -6,7 +6,7 @@ namespace ZealPHP\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use ZealPHP\App;
-use ZealPHP\Legacy\FastCgiClient;
+use ZealPHP\CGI\FastCgiClient;
 use ZealPHP\RequestContext;
 
 /**
